+++ src/resources.rs
@@ -3,8 +3,8 @@ use opengl_graphics::Texture;
 use font::*;
 
 
-pub const SPRITE_WIDTH:  f64 = 10.0;
-pub const SPRITE_HEIGHT: f64 = 10.0;
+pub const SPRITE_WIDTH:  u32 = 10;
+pub const SPRITE_HEIGHT: u32 = 10;
 
 pub struct Resources {
   pub big_font:      Font,
